requirements.txt
django==4.2.4
djangorestframework==3.14.0
mysqlclient==2.2.0
django-nested-inline==0.4.6
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
django-filter==23.2
django-rest-knox==4.2.0
django-cors-headers==4.2.0
